+++ keopscore/keopscore/__init__.py
@@ -14,6 +14,7 @@ with open(os.path.join(here, "keops_version"), encoding="utf-8") as v:
 # from keopscore.config.base_config import set_build_folder, get_build_folder  ## pas de setteurs et getters dans les fichiers
 from keopscore.config import *
 from keopscore.utils.code_gen_utils import clean_keops
+from keopscore.utils.code_gen_utils import check_health
 
 # flags for debugging :
 # prints information about atomic operations during code building
